requirements.txt
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.26.0
arch>=6.2.0
matplotlib>=3.8.0
plotly>=5.18.0
pyyaml>=6.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain>=0.1.0
shap>=0.44.0
lime>=0.2.0
aiohttp>=3.9.0
